Move RestClient into azure runtime
