+++ azure-client-runtime/src/main/java/com/microsoft/azure/AzureEnvironment.java
@@ -7,8 +7,6 @@
 
 package com.microsoft.azure;
 
-import com.microsoft.rest.RestClient;
-
 /**
  * An instance of this class describes an environment in Azure.
  */
@@ -86,8 +84,8 @@ public final class AzureEnvironment {
      *
      * @return a builder for the rest client.
      */
-    public RestClient.Builder newRestClientBuilder() {
-        return new AzureRestClient.Builder()
+    public RestClient.Builder.Buildable newRestClientBuilder() {
+        return new RestClient.Builder()
                 .withDefaultBaseUrl(this)
                 .withInterceptor(new RequestIdHeaderInterceptor());
     }
