+++ client-runtime/src/test/java/com/microsoft/rest/RetryHandlerTests.java
@@ -9,18 +9,17 @@ package com.microsoft.rest;
 
 import com.microsoft.rest.retry.RetryHandler;
 
-import com.microsoft.rest.serializer.JacksonMapperAdapter;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.io.IOException;
+
 import okhttp3.Interceptor;
 import okhttp3.OkHttpClient;
 import okhttp3.Protocol;
 import okhttp3.Request;
 import okhttp3.Response;
 import retrofit2.Retrofit;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-import java.io.IOException;
 
 public class RetryHandlerTests {
     @Test
