+++ client-runtime/src/test/java/com/microsoft/rest/ServiceClientTests.java
@@ -7,10 +7,8 @@
 
 package com.microsoft.rest;
 
-import com.microsoft.rest.serializer.JacksonMapperAdapter;
 import org.junit.Assert;
 import org.junit.Test;
-import retrofit2.Retrofit;
 
 import java.io.IOException;
 
@@ -19,6 +17,7 @@ import okhttp3.OkHttpClient;
 import okhttp3.Protocol;
 import okhttp3.Request;
 import okhttp3.Response;
+import retrofit2.Retrofit;
 
 public class ServiceClientTests {
     @Test
